requirements.txt
asgiref==3.4.1
Django==3.2.9
django-extensions==3.1.5
django-webpack-loader==1.4.1
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
